Add LibAlkimia[5]_INCLUDE_DIR to fix build issues with static libraries
Summary:
In cases where libraries are defined without library dependencies,
such as when creating static libraries, by default the location of the
alkimia include files is not added for compilation, resulting in compile
errors. This can be fixed by adding LibAlkimia[5]_INCLUDE_DIR to the list
of specified include directories. The camel case style of the the first
part of this variable name has been choosen to match the name given to
find_package() for finding alkimia.

For compatibility with old code the deprectated variable LIBALKIMIA_INCLUDE_DIR
has been reintroduced.

Signed-off-by: Ralf Habacker <ralf.habacker@freenet.de>

Differential Revision: https://phabricator.kde.org/D11539