Life cycle of the instance of the AlkWebView/AlkWebPage class improved
The creation of instances of the AlkWebView and AlkWebPage classes has
been moved out of the AlkOnlineProfileManager in order to have more
control over the life cycle.

The AlkOnlineProfileManager class now only contains pointers to
corresponding instances that must be set externally.

With this commit, the warning reported by QWebEngine

  Release of profile requested but WebEnginePage still not deleted.
  Expect troubles !

should be fixed.

BUG:485662
FIXED-IN:8.2